Allow for an app to start with low (or zero) brightness if the previously-running app set it that way.
This is useful if we want to use a long/short presses to launch different apps (where the first one starts launching on a short press, but then checks for a long press and potentially loading a second app instead before turning the backlight on).

However, this means a reset() issued over BLE will no longer turn on the backlight automatically.